fix(monitor): 清掉 golangci-lint 报出的既存问题
v1.18.2 的 CI 在 golangci-lint 步骤失败。这 8 条报错都是 v1.18.1 起就存在的
既存问题，与上一次测试修复无关（报错文件中 6 个未被改动；
cloudwatch_logs_parse_newapi.go 的 rule 声明在 v1.18.1 原文即如此，
只是行号被推移）。本次全部修掉，使该步骤可以通过。

- logchain_investigation.go: 去掉 if-return 之后的 else，模型/分组校验改为
  平铺写法（revive indent-error-flow），语义不变；
- cloudwatch_logs_evidence.go、logchain_investigation_test.go:
  错误判定改用 errors.As（errorlint）。这不只是过 lint——原先的类型断言
  遇到 fmt.Errorf("%w") 包装过的错误会误判成 malformed / 非预期错误；
- cloudwatch_shadow_runner.go: 去掉 ^uint64(0)>>1 外层冗余转换（unconvert）；
- cloudwatch_logs_parse_newapi.go: rule/ok 改为先声明，去掉立即被覆盖的初值
  （staticcheck SA4006）；
- cloudwatch_logs_parse_common.go: 删除 cwBoundedLabel（unused）。它在
  v1.18.1 引入后从未被任何代码调用，同文件已有在用的 cwBusinessLabel，
  另有 cwBoundedOpaque / cwInvestigationScopeLabel 承担同类边界校验；
  如需恢复可从 git 历史取回；
- source_lifecycle.go: gofmt 结构体字段对齐。git diff -w 为空，无语义变化。

验证（均真实执行并成功退出）：golangci-lint v1.64.8（GOOS=linux，与 CI 同版本
同配置）真实问题 0，仅余 3 处本机 core.autocrlf=true 造成的 gofmt 假报
（剥除 CR 后 gofmt 判定干净，CI 侧不出现）；go vet ./...（linux/amd64）通过；
race 分片 a-f / G-N / O-S / T-Z 全部 0 失败；check-race-shards 通过且
1707 个测试分布不变。

本机无 python3、缺 gcc 与 go1.26 toolchain，因此 test job 的 python 步骤、
非 monitor 包 race 测试、govulncheck 仍未在本地执行，由 CI 覆盖。

+++ monitor/cloudwatch_shadow_runner.go
@@ -89,7 +89,7 @@ func (m *Monitor) queryCloudWatchShadowRange(ctx context.Context, parser *cloudW
 	query, err := m.cloudWatchLogs.insights(ctx, cloudWatchInsightsRequest{
 		Kind: kind, From: from, To: to, Limit: cloudWatchShadowQueryLimit,
 	})
-	if query.BytesScanned <= uint64(^uint64(0)>>1) {
+	if query.BytesScanned <= ^uint64(0)>>1 {
 		*bytes += int64(query.BytesScanned)
 	}
 	if err != nil {
